+++ lib/generators/templates/lokalise_rails_config.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'lokalise_rails'
-
 LokaliseRails.config do |c|
   # These are mandatory options that you must set before running rake tasks:
   # c.api_token = ENV['LOKALISE_API_TOKEN']