+++ spec/lib/tasks/export_task_spec.rb
@@ -1,19 +1,23 @@
 # frozen_string_literal: true
 
-RSpec.describe LokaliseRails do
-  let(:fake_exporter) { LokaliseRails::TaskDefinition::Exporter }
+RSpec.describe "Export Rake task" do
+  let(:described_klass) { LokaliseRails::TaskDefinition::Exporter }
+  let(:exporter_object) { described_klass.new }
+  let(:top_object) { LokaliseRails }
 
   it 'halts when the API key is not set' do
-    allow(described_class).to receive(:api_token).and_return(nil)
+    allow(top_object).to receive(:api_token).and_return(nil)
 
     expect(export_executor).to raise_error(SystemExit, /API token is not set/i)
-    expect(described_class).to have_received(:api_token)
+    expect(top_object).to have_received(:api_token)
   end
 
   it 'halts when the project ID is not set' do
-    allow_project_id nil do
-      expect(export_executor).to raise_error(SystemExit, /ID is not set/i)
-    end
+    allow(top_object).to receive(:project_id).and_return(nil)
+
+    expect(export_executor).to raise_error(SystemExit, /ID is not set/i)
+
+    expect(top_object).to have_received(:project_id)
   end
 
   context 'with two translation files' do
@@ -30,26 +34,30 @@ RSpec.describe LokaliseRails do
     end
 
     it 'handles too many requests' do
-      allow_project_id '672198945b7d72fc048021.15940510'
-      allow(described_class).to receive(:max_retries_export).and_return(2)
-      allow(fake_exporter).to receive(:sleep).and_return(0)
+      allow_task_def_instance described_klass, exporter_object
+      allow_project_id exporter_object, '672198945b7d72fc048021.15940510'
+
+      allow(exporter_object.global_opts).to receive(:max_retries_export).and_return(2)
+      allow(exporter_object).to receive(:sleep).and_return(0)
 
       fake_client = double
       allow(fake_client).to receive(:upload_file).and_raise(Lokalise::Error::TooManyRequests)
-      allow(fake_exporter).to receive(:api_client).and_return(fake_client)
+      allow(exporter_object).to receive(:api_client).and_return(fake_client)
 
       expect(export_executor).to raise_error(SystemExit, /Gave up after 2 retries/i)
 
-      expect(fake_exporter).to have_received(:sleep).exactly(2).times
-      expect(described_class).to have_received(:max_retries_export).exactly(1).time
-      expect(fake_exporter).to have_received(:api_client).exactly(3).times
+      expect(exporter_object).to have_received(:sleep).exactly(2).times
+      expect(exporter_object.global_opts).to have_received(:max_retries_export).exactly(1).time
+      expect(exporter_object).to have_received(:api_client).exactly(3).times
       expect(fake_client).to have_received(:upload_file).exactly(3).times
     end
 
     it 'runs export rake task properly' do
       fake_client = double
-      allow(fake_client).to receive(:upload_file).and_return(true)
-      allow(fake_exporter).to receive(:api_client).and_return(fake_client)
+      allow_task_def_instance described_klass, exporter_object
+
+      allow(fake_client).to receive(:upload_file).with(any_args).and_return(true)
+      allow(exporter_object).to receive(:api_client).and_return(fake_client)
 
       expect(export_executor).to output(/complete!/).to_stdout
 
@@ -58,7 +66,8 @@ RSpec.describe LokaliseRails do
 
     describe '.export!' do
       it 're-raises export errors' do
-        allow_project_id
+        allow_task_def_instance described_klass, exporter_object
+        allow_project_id exporter_object
 
         VCR.use_cassette('upload_files_error') do
           expect(export_executor).to raise_error(SystemExit, /Unknown `lang_iso`/)