+++ spec/support/spec_addons.rb
@@ -1,11 +1,15 @@
 # frozen_string_literal: true
 
 module SpecAddons
-  def allow_project_id(value = '189934715f57a162257d74.88352370')
-    allow(LokaliseRails).to receive(:project_id).and_return(value)
+  def allow_project_id(obj, value = '189934715f57a162257d74.88352370')
+    allow(obj.global_opts).to receive(:project_id).and_return(value)
     return unless block_given?
 
     yield
-    expect(LokaliseRails).to have_received(:project_id)
+    expect(obj.global_opts).to have_received(:project_id)
+  end
+
+  def allow_task_def_instance(klass, instance)
+    allow(klass).to receive(:new).and_return(instance)
   end
 end